% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account_details.R
\name{get_account_details}
\alias{get_account_details}
\title{Get Uptime Robot account details}
\usage{
get_account_details(api_key)
}
\arguments{
\item{api_key}{The api_key}
}
\description{
Returns the account details
}
