% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monitors.R
\name{get_monitors}
\alias{get_monitors}
\title{Get monitors}
\usage{
get_monitors(
  api_key,
  logs = FALSE,
  all_uptime_ratio = FALSE,
  all_time_uptime_durations = FALSE,
  custom_uptime_ratios = NULL,
  response_times = FALSE
)
}
\arguments{
\item{api_key}{The api_key}

\item{logs}{Should the logs be returned (default FALSE)}

\item{all_uptime_ratio}{Uptime ratio since monitoring began (default FALSE).}

\item{all_time_uptime_durations}{All time uptime durations (default FALSE).}

\item{custom_uptime_ratios}{Defines the number of days to calculate the uptime ratio(s) for.
Ex: custom_uptime_ratios=7-30-45 to get the uptime ratios for those periods.}

\item{response_times}{Logical default \code{FALSE}.}
}
\description{
Get monitors
}
\details{
The argument order may change. So use named arguments. I also want to
make this function to always return a tibble with the same number of columns. However,
the api itself isn't really consistent and depends on the arguments passed.
}
\seealso{
https://uptimerobot.com/api
}
